requirements.txt
langchain>=0.0.267
openai>=0.27.8
chromadb>=0.4.13
python-dotenv>=1.0.0
pandas>=2.0.3
jupyter>=1.0.0
ipykernel>=6.25.1
numpy>=1.24.3
matplotlib>=3.7.2
